+++ src/GRA.Data/Model/ReportCriterion.cs
@@ -11,6 +11,9 @@ namespace GRA.Data.Model
         public int? SystemId { get; set; }
         public int? BranchId { get; set; }
         public int? ProgramId { get; set; }
+        public int? SchoolDistrictId { get; set; }
+        public int? SchoolId { get; set; }
+
         public bool Favorite { get; set; }
         [MaxLength(255)]
         public string Name { get; set; }