+++ src/GRA.Web/Startup.cs
@@ -201,7 +201,7 @@ namespace GRA.Web
             services.AddScoped<Domain.Report.BadgeTopScoresReport>();
             services.AddScoped<Domain.Report.CurrentStatusByProgramReport>();
             services.AddScoped<Domain.Report.CurrentStatusReport>();
-            services.AddScoped<Domain.Report.RegisteredBySchoolReport>();
+            services.AddScoped<Domain.Report.RegistrationsAchieversBySchoolReport>();
             services.AddScoped<Domain.Report.RegistrationsAchieversReport>();
             services.AddScoped<Domain.Report.ParticipantProgressReport>();
             services.AddScoped<Domain.Report.TopScoresReport>();